added persistend to save exist sign

+++ src/main/java/com/example/mobilephonebooking/services/BookingPhone.java
@@ -1,10 +1,9 @@
 package com.example.mobilephonebooking.services;
 
 import com.example.mobilephonebooking.model.Phone;
-import reactor.core.publisher.Mono;
 
 import java.util.Optional;
 
 public interface BookingPhone {
-    Optional<Phone> book(int id);
+    Optional<Phone> book(long id);
 }
